+++ store/src/Haskoin/Store/Stats.hs
@@ -4,7 +4,6 @@ module Haskoin.Store.Stats
     , withStats
     , createStatDist
     , addStatTime
-    , addStatItems
     , addClientError
     , addServerError
     , addStatQuery
@@ -52,7 +51,6 @@ data StatData =
     StatData
     {
         statTimes        :: ![Int64],
-        statItems        :: !Int64,
         statQueries      :: !Int64,
         statClientErrors :: !Int64,
         statServerErrors :: !Int64
@@ -62,7 +60,6 @@ data StatDist =
     StatDist
     {
         distQueue        :: !(TQueue Int64),
-        distItems        :: !(TVar Int64),
         distQueries      :: !(TVar Int64),
         distClientErrors :: !(TVar Int64),
         distServerErrors :: !(TVar Int64)
@@ -71,13 +68,11 @@ data StatDist =
 createStatDist :: MonadIO m => Text -> Store -> m StatDist
 createStatDist t store = liftIO $ do
     q <- newTQueueIO
-    items <- newTVarIO 0
     queries <- newTVarIO 0
     client_errors <- newTVarIO 0
     server_errors <- newTVarIO 0
     let metrics = HashMap.fromList
             [ (t <> ".query_count",   Counter . statQueries)
-            , (t <> ".item_count",    Counter . statItems)
             , (t <> ".errors.client", Counter . statClientErrors)
             , (t <> ".errors.server", Counter . statServerErrors)
             , (t <> ".mean_ms",       Gauge . mean . statTimes)
@@ -85,11 +80,10 @@ createStatDist t store = liftIO $ do
             , (t <> ".max_ms",        Gauge . maxi . statTimes)
             , (t <> ".min_ms",        Gauge . mini . statTimes)
             , (t <> ".p90max_ms",     Gauge . p90max . statTimes)
-            , (t <> ".p90min_ms",     Gauge . p90min . statTimes)
             , (t <> ".p90avg_ms",     Gauge . p90avg . statTimes)
             , (t <> ".var_ms",        Gauge . var . statTimes)
             ]
-    let sd = StatDist q items queries client_errors server_errors
+    let sd = StatDist q queries client_errors server_errors
     registerGroup metrics (flush sd) store
     return sd
 
@@ -99,10 +93,6 @@ toDouble = fromIntegral
 addStatTime :: MonadIO m => StatDist -> Int64 -> m ()
 addStatTime q =
     liftIO . atomically . writeTQueue (distQueue q)
-
-addStatItems :: MonadIO m => StatDist -> Int64 -> m ()
-addStatItems q =
-    liftIO . atomically . modifyTVar (distItems q) . (+)
 
 addStatQuery :: MonadIO m => StatDist -> m ()
 addStatQuery q =
@@ -117,13 +107,12 @@ addServerError q =
     liftIO . atomically $ modifyTVar (distServerErrors q) (+1)
 
 flush :: MonadIO m => StatDist -> m StatData
-flush (StatDist q i n c s) = atomically $ do
+flush (StatDist q n c s) = atomically $ do
     ts <- flushTQueue q
-    is <- readTVar i
     qs <- readTVar n
     ce <- readTVar c
     se <- readTVar s
-    return $ StatData ts is qs ce se
+    return $ StatData ts qs ce se
 
 average :: Fractional a => L.Fold a a
 average = (/) <$> L.sum <*> L.genericLength
@@ -150,16 +139,6 @@ p90max ls =
         h:_ -> h
   where
     sorted = sortBy (comparing Down) ls
-    len = length sorted
-    chopped = drop (length sorted * 1 `div` 10) sorted
-
-p90min :: [Int64] -> Int64
-p90min ls =
-    case chopped of
-        []  -> 0
-        h:_ -> h
-  where
-    sorted = sort ls
     len = length sorted
     chopped = drop (length sorted * 1 `div` 10) sorted